+++ chaosmagpy/__init__.py
@@ -4,7 +4,7 @@ __all__ = [
     "basicConfig", "synth_values"
 ]
 
-__version__ = "0.9-dev"
+__version__ = "0.9"
 
 from .chaos import (
     CHAOS,