+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog
 Version 0.9-dev
 ---------------
 | **Date:** April 1, 2022
-| **Release:** v0.9-dev
+| **Release:** v0.9
 | **Version of CHAOS:** CHAOS-7.10 (0710)
 
 Features